Sufijos y consultas para modelo

+++ siga-core/src/main/java/org/itcgae/siga/commons/constants/SigaConstants.java
@@ -255,20 +255,20 @@ public class SigaConstants {
 	}
 	
 	public static enum OBJETIVO {
-		DESTINATARIOS(new Long("1"), "DESTINATARIOS"),
-		MULTIDOCUMENTO(new Long("2"), "MULTIDOCUMENTO"),
-		CONDICIONAL(new Long("3"), "CONDICIONAL"),
-		DATOS(new Long("4"), "DATOS");
+		DESTINATARIOS(new Short("1"), "DESTINATARIOS"),
+		MULTIDOCUMENTO(new Short("2"), "MULTIDOCUMENTO"),
+		CONDICIONAL(new Short("3"), "CONDICIONAL"),
+		DATOS(new Short("4"), "DATOS");
 		
-		private final Long codigo;
+		private final Short codigo;
 		private final String descripcion;
 
-		OBJETIVO(Long codigo, String descripcion) {
+		OBJETIVO(Short codigo, String descripcion) {
 			this.codigo = codigo;
 			this.descripcion = descripcion;
 		}
 
-		public Long getCodigo() {
+		public Short getCodigo() {
 			return codigo;
 		}
